비밀번호 확인 기능 추가

+++ Schedule-Project/src/main/java/org/example/scheduleproject/service/ScheduleService.java
@@ -1,8 +1,10 @@
 package org.example.scheduleproject.service;
 
+import lombok.RequiredArgsConstructor;
 import org.example.scheduleproject.dto.RequestScheduleWithUserDto;
 import org.example.scheduleproject.dto.ResponseScheduleDto;
 import org.example.scheduleproject.dto.UpdateTodoList;
+import org.example.scheduleproject.dto.UserDto;
 import org.example.scheduleproject.repository.ScheduleRepository;
 import org.example.scheduleproject.repository.UserRepository;
 import org.springframework.stereotype.Service;
@@ -10,19 +12,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
     private final UserRepository userRepository;
 
-    public ScheduleService(ScheduleRepository scheduleRepository, UserRepository userRepository) {
-        this.scheduleRepository = scheduleRepository;
-        this.userRepository = userRepository;
-    }
 
     @Transactional
     public UUID save(RequestScheduleWithUserDto requestScheduleWithUserDto) {
@@ -35,7 +33,7 @@ public class ScheduleService {
     }
 
 
-    public Optional<ResponseScheduleDto> getTodoList(UUID scheduleId) {
+    public ResponseScheduleDto getTodoList(UUID scheduleId) {
         return scheduleRepository.findScheduleById(scheduleId);
     }
 
@@ -43,12 +41,30 @@ public class ScheduleService {
         return scheduleRepository.findAllSchedule();
     }
 
-    public void deleteSchedule(UUID scheduleId) {
-        scheduleRepository.deleteScheduleById(scheduleId);
+    @Transactional
+    public void deleteSchedule(UUID scheduleId, String password) {
+        UserDto userById = findUserById(scheduleId);
+        if (password.equals(userById.getPassword())) {
+            scheduleRepository.deleteScheduleById(scheduleId);
+            userRepository.deleteUser(scheduleId);
+            return;
+        }
+        throw new SecurityException("비밀번호가 일치하지 않습니다.");
     }
 
     public UUID updateSchedule(UUID scheduleId, UpdateTodoList updateTodoList) {
-        return scheduleRepository.update(scheduleId, updateTodoList);
+        UserDto userById = findUserById(scheduleId);
+
+        if (updateTodoList.getPassword().equals(userById.getPassword())) {
+            return scheduleRepository.update(scheduleId, updateTodoList);
+        }
+
+        throw new SecurityException("비밀번호가 일치하지 않습니다.");
+    }
+
+    private UserDto findUserById(UUID scheduleId) {
+        ResponseScheduleDto findSchedule = scheduleRepository.findScheduleById(scheduleId);
+        return userRepository.findUserPasswordById(findSchedule.getUserId());
     }
 
 }
